requirements.txt
openai>=1.0.0
aiohttp>=3.8.0
asyncio
dataclasses
typing-extensions
python-dotenv>=1.0.0
python-dotenv
duckduckgo-search
pytest
pytest-asyncio
"typer[all]"
websockets
httpx
pytest-httpx
